requirements.txt
fastapi
pandas
fastapi
uvicorn
python-dotenv
python-multipart